removed post id from user posts array

+++ src/controllers/post.controller.ts
@@ -4,7 +4,10 @@ import { POSTS } from "../constant.js";
 import { ApiRes } from "../utils/response.js";
 import { Post } from "../models/post.model.js";
 import { asyncGuard } from "../utils/asyncGuard.js";
-import { uploadFileOneCloud } from "../helpers/cloudinary.helper.js";
+import {
+  deleteFileFromCloud,
+  uploadFileOneCloud,
+} from "../helpers/cloudinary.helper.js";
 
 const getPosts = asyncGuard(async (req, res) => {
   const limit = Math.min(parseInt(req.query.limit as string) || 10);
@@ -55,12 +58,25 @@ const createPost = asyncGuard(async (req, res) => {
 
 const deletePost = asyncGuard(async (req, res) => {
   const _id = req.params.postId;
+  const userId = req.user?._id;
 
   if (!isValidObjectId(_id)) {
     return res.status(400).json(new ApiRes(400, "Valid post id is required"));
   }
 
+  const post = await Post.findById(_id);
+
+  if (!post) {
+    return res.status(400).json(new ApiRes(400, "Post not found"));
+  }
+
+  if (post.postPublicId && post.postType) {
+    await deleteFileFromCloud(post.postPublicId, post.postType);
+  }
+
+  await User.findByIdAndUpdate(userId, { $pull: { posts: _id } });
   await Post.findByIdAndDelete(_id);
+
   return res.status(200).json(new ApiRes(200, "Post deleted"));
 });
